requirements.txt
requests==2.31.0
feedparser==6.0.11
google-generativeai==0.8.3
jinja2==3.1.4
python-dotenv==1.0.1
beautifulsoup4==4.12.3
lxml==5.3.0
